project4/lambda/requirements.txt
boto3
pillow

# Removing torch/torchvision from here to install via explicit pip command with --find-links/index-url in Dockerfile
